pyproject.toml
[tool.poetry]
name = "example"
version = "0.1.0"
description = "Very interesting example, yeah!"
authors = ["Nikita Sivakov <sivakov512@gmail.com>"]

[tool.poetry.scripts]
example = 'cli:cli'

[tool.poetry.dependencies]
grpcio = "^1.41.1"
protobuf = "^3.19.1"
python = "^3.9"
click = "^8.0.3"
grpcio-tools = "^1.41.1"
asyncpg = "^0.24.0"
asyncpg-engine = "^0.2.0"

[tool.poetry.dev-dependencies]
black = "^21.10b0"
flake8 = "^4.0.1"
flake8-black = "^0.2.3"
flake8-debugger = "^4.0.0"
flake8-eradicate = "^1.2.0"
flake8-isort = "^4.1.1"
flake8-print = "^4.0.0"
flake8-pytest-style = "^1.5.1"
flake8-quotes = "^3.3.1"
grpc-stubs = "^1.24.7"
isort = "^5.10.0"
mypy = "^0.910"
mypy-protobuf = "^3.0.0"
pytest = "^6.2.5"
pytest-asyncio = "^0.16.0"
pytest-deadfixtures = "^2.2.1"
types-protobuf = "^3.18.0"
pdbpp = "^0.10.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
combine_as_imports = true
include_trailing_comma = true
line_length = 99
multi_line_output = 3
use_parentheses = true

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
explicit_package_bases = true
ignore_missing_imports = true
implicit_reexport = false
namespace_packages = true
no_implicit_optional = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["proto.*"]
ignore_errors = true

[tool.black]
line-length = 100
